refactor: 안 쓰이는 레거시 코드 완벽 제거
