+++ README.md
@@ -36,5 +36,5 @@
 - [x] 테스트 system.out 수정
 - [ ] 인스턴스 변수 this 활용에 기준 잡기
 - [x] 에러 처리에 있어 jspView 중복 제거
-- [ ] 안 쓰이는 레거시 코드 완벽 제거
+- [x] 안 쓰이는 레거시 코드 완벽 제거
 - [ ] JsonView에서 여러 파라미터의 요청에 대해 응답 받아볼 수 있는 컨트롤러 메서드 추가
