+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -1,6 +1,5 @@
-package com.techcourse.controller.tobe;
+package com.techcourse.controller;
 
-import com.techcourse.controller.UserSession;
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
