+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller.tobe;
+package com.techcourse.controller;
 
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
