+++ app/src/main/java/com/techcourse/controller/TestController.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller.tobe;
+package com.techcourse.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
