--- app/src/main/java/com/techcourse/controller/asis/LogoutController.java
@@ -1,17 +1,0 @@
-package com.techcourse.controller.asis;
-
-import com.techcourse.controller.UserSession;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
-import nextstep.mvc.controller.asis.Controller;
-
-public class LogoutController implements Controller {
-
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse res) {
-        final HttpSession session = req.getSession();
-        session.removeAttribute(UserSession.SESSION_KEY);
-        return "redirect:/";
-    }
-}
