--- app/src/main/java/com/techcourse/controller/asis/RegisterController.java
@@ -1,21 +1,0 @@
-package com.techcourse.controller.asis;
-
-import com.techcourse.domain.User;
-import com.techcourse.repository.InMemoryUserRepository;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
-
-public class RegisterController implements Controller {
-
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse res) {
-        final User user = new User(
-                req.getParameter("account"),
-                req.getParameter("password"),
-                req.getParameter("email"));
-        InMemoryUserRepository.save(user);
-
-        return "redirect:/index.jsp";
-    }
-}
