--- mvc/src/main/java/nextstep/mvc/controller/asis/ControllerHandlerAdapter.java
@@ -1,22 +1,0 @@
-package nextstep.mvc.controller.asis;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-
-public class ControllerHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) {
-        final Controller controller = (Controller) handler;
-        final String view = controller.execute(request, response);
-        return new ModelAndView(new JspView(view));
-    }
-}
